Removing unused imports
